reafactor: optimization changes
