+++ src/components/carousel/carousel.constants.tsx
@@ -1,0 +1,29 @@
+import avatar1 from '../../../public/avatar-1.svg'
+import { SlideData } from './carousel.types'
+
+export const slidesData: SlideData[] = [
+  {
+    avatar: avatar1,
+    description:
+      'Como doadora do IPAS, testemunhei de perto a transformação que ele é capaz de realizar. Ver as crianças, os jovens e os adultos vulneráveis sendo beneficiados pelos projetos de educação, cultura e assistência social é verdadeiramente inspirador. Estou orgulhosa de fazer parte desse movimento que combate a desigualdade e promove oportunidades para todos.',
+    name: '— Mariana Souza, doadora do IPAS.',
+  },
+  {
+    avatar: avatar1,
+    description:
+      'Como doadora do IPAS, testemunhei de perto a transformação que ele é capaz de realizar.',
+    name: '— Mariana Souza 2',
+  },
+  {
+    avatar: avatar1,
+    description:
+      'Ver as crianças, os jovens e os adultos vulneráveis sendo beneficiados pelos projetos de educação, cultura e assistência social é verdadeiramente inspirador.',
+    name: '— Mariana Souza 3',
+  },
+  {
+    avatar: avatar1,
+    description:
+      'Estou orgulhosa de fazer parte desse movimento que combate a desigualdade e promove oportunidades para todos.',
+    name: '— Mariana Souza 4',
+  },
+]
