+++ src/components/carousel/carousel.types.ts
@@ -1,0 +1,5 @@
+export type SlideData = {
+  avatar: string
+  description: string
+  name: string
+}
