+++ src/components/carousel/index.tsx
@@ -1,1 +1,2 @@
 export { Carousel } from './carousel'
+export type { SlideData } from './carousel.types'
